+++ psef/models/course.py
@@ -3,6 +3,7 @@
 SPDX-License-Identifier: AGPL-3.0-only
 """
 import copy
+import enum
 import uuid
 import typing as t
 
@@ -11,18 +12,20 @@ from sqlalchemy.orm import selectinload
 from typing_extensions import TypedDict
 
 import psef
+import cg_enum
 from cg_dt_utils import DatetimeWithTimezone
 from cg_typing_extensions import make_typed_dict_extender
 from cg_sqlalchemy_helpers import mixins, expression
 
 from . import Base, MyQuery, db
+from . import role as role_models
+from . import user as user_models
+from . import work as work_models
 from .. import auth
-from .role import CourseRole
-from .user import User
-from .work import Work
 from ..helpers import NotEqualMixin, jsonify_options
 from .assignment import Assignment
 from .link_tables import user_course
+from ..permissions import CoursePermission
 
 logger = structlog.get_logger()
 
@@ -59,7 +62,9 @@ class CourseRegistrationLink(Base, mixins.UUIDMixin, mixins.TimestampMixin):
         innerjoin=True,
     )
     course_role = db.relationship(
-        lambda: CourseRole, foreign_keys=course_role_id, innerjoin=True
+        lambda: role_models.CourseRole,
+        foreign_keys=course_role_id,
+        innerjoin=True
     )
 
     allow_register = db.Column(
@@ -122,7 +127,16 @@ class CourseSnippet(Base):
         }
 
 
-class Course(NotEqualMixin, Base):
+@enum.unique
+class CourseState(cg_enum.CGEnum):
+    """Describes in what state a :class:`.Course` is.
+    """
+    visible = 'visible'
+    archived = 'archived'
+    deleted = 'deleted'
+
+
+class Course(NotEqualMixin, Base, mixins.TimestampMixin, mixins.IdMixin):
     """This class describes a course.
 
     A course can hold a collection of :class:`.Assignment` objects.
@@ -132,13 +146,14 @@ class Course(NotEqualMixin, Base):
     :param lti_provider: The LTI provider
     """
     __tablename__ = "Course"
-    id = db.Column('id', db.Integer, primary_key=True)
     name = db.Column('name', db.Unicode, nullable=False)
 
-    created_at = db.Column(
-        db.TIMESTAMP(timezone=True),
-        default=DatetimeWithTimezone.utcnow,
+    state = db.Column(
+        'state',
+        cg_enum.CGDbEnum(CourseState),
         nullable=False,
+        default=CourseState.visible,
+        server_default=CourseState.visible.name,
     )
 
     course_lti_provider = db.relationship(
@@ -169,6 +184,7 @@ class Course(NotEqualMixin, Base):
         cascade='all,delete',
         uselist=True,
         order_by=lambda: psef.models.GroupSet.created_at,
+        lazy='select',
     )
 
     snippets = db.relationship(
@@ -186,6 +202,7 @@ class Course(NotEqualMixin, Base):
         cascade='all,delete',
         uselist=True,
         order_by=lambda: CourseRegistrationLink.created_at,
+        lazy='select',
     )
 
     assignments = db.relationship(
@@ -193,6 +210,7 @@ class Course(NotEqualMixin, Base):
         back_populates="course",
         cascade='all,delete',
         uselist=True,
+        lazy='select',
     )
 
     class AsJSON(TypedDict, total=True):
@@ -213,6 +231,8 @@ class Course(NotEqualMixin, Base):
         #: The lti provider that manages this course, if ``null`` this is not a
         #: LTI course.
         lti_provider: t.Optional['psef.models.LTIProviderBase']
+        #: The state this course is in.
+        state: CourseState
 
     class AsExtendedJSON(AsJSON, total=True):
         """The way this class will be represented in extended JSON.
@@ -245,6 +265,7 @@ class Course(NotEqualMixin, Base):
             'is_lti': self.is_lti,
             'virtual': self.virtual,
             'lti_provider': self.lti_provider,
+            'state': self.state,
         }
         if jsonify_options.get_options().add_role_to_course:
             user = psef.current_user
@@ -308,8 +329,9 @@ class Course(NotEqualMixin, Base):
         if virtual:
             return self
 
-        for role_name, perms in CourseRole.get_default_course_roles().items():
-            CourseRole(
+        for role_name, perms in role_models.CourseRole.get_default_course_roles(
+        ).items():
+            role_models.CourseRole(
                 name=role_name, course=self, _permissions=perms, hidden=False
             )
 
@@ -318,10 +340,10 @@ class Course(NotEqualMixin, Base):
         admin_username = psef.current_app.config['ADMIN_USER']
 
         if admin_username is not None:
-            admin_user = User.query.filter_by(
+            admin_user = user_models.User.query.filter_by(
                 username=admin_username,
             ).one_or_none()
-            admin_role = CourseRole.get_admin_role(self)
+            admin_role = role_models.CourseRole.get_admin_role(self)
 
             if admin_user is None:
                 logger.error(
@@ -352,7 +374,7 @@ class Course(NotEqualMixin, Base):
     def __eq__(self, other: object) -> bool:
         """Check if two courses are equal.
 
-        >>> CourseRole.get_default_course_roles = lambda: {}
+        >>> role_models.CourseRole.get_default_course_roles = lambda: {}
         >>> c1 = Course()
         >>> c1.id = 5
         >>> c2 = Course()
@@ -396,24 +418,45 @@ class Course(NotEqualMixin, Base):
     def get_assignments(self) -> MyQuery['Assignment']:
         return Assignment.query.filter(Assignment.course == self)
 
-    def get_all_users_in_course(self, *, include_test_students: bool
-                                ) -> MyQuery['t.Tuple[User, CourseRole]']:
+    def get_all_users_in_course(
+        self,
+        *,
+        include_test_students: bool,
+        with_permission: CoursePermission = None
+    ) -> MyQuery['t.Tuple[user_models.User, role_models.CourseRole]']:
         """Get a query that returns all users in the current course and their
             role.
 
         :returns: A query that contains all users in the current course and
             their role.
         """
-        res = db.session.query(User, CourseRole).join(
+        CourseRole = role_models.CourseRole
+
+        join_conds = [CourseRole.id == user_course.c.course_id]
+        if with_permission is not None:
+            join_conds.append(
+                CourseRole.id.in_(
+                    CourseRole.get_roles_with_permission(
+                        with_permission,
+                    ).filter(CourseRole.course_id == self.id).with_entities(
+                        CourseRole.id
+                    )
+                )
+            )
+
+        res = db.session.query(user_models.User, CourseRole).join(
             user_course,
-            user_course.c.user_id == User.id,
+            user_course.c.user_id == user_models.User.id,
         ).join(
             CourseRole,
-            CourseRole.id == user_course.c.course_id,
-        ).filter(CourseRole.course_id == self.id, User.virtual.isnot(True))
+            expression.and_(*join_conds),
+        ).filter(
+            CourseRole.course_id == self.id,
+            user_models.User.virtual.isnot(True),
+        )
 
         if not include_test_students:
-            res = res.filter(~User.is_test_student)
+            res = res.filter(~user_models.User.is_test_student)
 
         return res
 
@@ -443,8 +486,9 @@ class Course(NotEqualMixin, Base):
         for child in copy.copy(tree.values):
             # This is done before we wrap single files to get better author
             # names.
-            work = Work(
-                assignment=assig, user=User.create_virtual_user(child.name)
+            work = work_models.Work(
+                assignment=assig,
+                user=user_models.User.create_virtual_user(child.name)
             )
 
             subdir: psef.files.ExtractFileTreeBase
@@ -460,7 +504,7 @@ class Course(NotEqualMixin, Base):
             work.add_file_tree(subdir)
         return self
 
-    def get_test_student(self) -> User:
+    def get_test_student(self) -> 'user_models.User':
         """Get the test student for this course. If no test student exists yet
         for this course, create a new one and return that.
 
@@ -468,17 +512,17 @@ class Course(NotEqualMixin, Base):
         """
 
         user = self.get_all_users_in_course(include_test_students=True).filter(
-            User.is_test_student,
-        ).from_self(User).first()
+            user_models.User.is_test_student,
+        ).from_self(user_models.User).first()
 
         if user is None:
-            role = CourseRole(
+            role = role_models.CourseRole(
                 name=f'Test_Student_Role__{uuid.uuid4()}',
                 course=self,
                 hidden=True,
             )
             db.session.add(role)
-            user = User.create_new_test_student()
+            user = user_models.User.create_new_test_student()
             user.courses[self.id] = role
             db.session.add(user)