+++ src/components/HomeGrid.vue
@@ -18,8 +18,8 @@
 
     <b-alert show v-if="showReleaseNote" variant="info">
         A new version of CodeGrade has been released:
-        <b>{{ UserConfig.release.version }}</b>.
-        {{ UserConfig.release.message }} You can check the entire
+        <b>{{ $userConfig.release.version }}</b>.
+        {{ $userConfig.release.message }} You can check the entire
         changelog <a href="https://docs.codegra.de/about/changelog.html"
                      target="_blank"
                      class="alert-link">here</a>.
@@ -47,12 +47,7 @@
                 <b-card-header :class="`text-${getColorPair(course.name).color}`"
                                :style="{ backgroundColor: `${getColorPair(course.name).background} !important` }">
                     <div style="display: flex">
-                        <div class="course-name">
-                            <b>{{ course.name }}</b>
-                            <i v-if="courseExtraDataToDisplay[course.id]">
-                                ({{ courseExtraDataToDisplay[course.id] }})
-                            </i>
-                        </div>
+                        <course-name :course="course" :bold="true" />
                         <router-link v-if="course.canManage"
                                      :to="manageCourseRoute(course)"
                                      v-b-popover.window.top.hover="'Manage course'"
@@ -113,7 +108,6 @@ import 'vue-awesome/icons/gear';
 import InfiniteLoading from 'vue-infinite-loading';
 
 import { hashString } from '@/utils';
-import { Counter } from '@/utils/counter';
 import { INITIAL_COURSES_AMOUNT } from '@/constants';
 
 import AssignmentState from './AssignmentState';
@@ -122,6 +116,7 @@ import Loader from './Loader';
 import LocalHeader from './LocalHeader';
 import CgLogo from './CgLogo';
 import AssignmentListItem from './Sidebar/AssignmentListItem';
+import CourseName from './CourseName';
 
 // We can't use the COLOR_PAIRS from constants.js because that one is slightly
 // different and because we use hashes to index this list that would change most
@@ -151,7 +146,7 @@ const COLOR_PAIRS = [
 
 // The amount of extra courses that should be loaded when we reach the end of
 // the infinite scroll list.
-const EXTRA_COURSES_AMOUNT = INITIAL_COURSES_AMOUNT / 2;
+const EXTRA_COURSES_AMOUNT = INITIAL_COURSES_AMOUNT / 3;
 
 export default {
     name: 'home-grid',
@@ -159,7 +154,6 @@ export default {
     data() {
         return {
             loadingCourses: true,
-            UserConfig,
             amountCoursesToShow: EXTRA_COURSES_AMOUNT,
             searchString: this.$route.query.filter || '',
             renderingMoreCourses: 0,
@@ -170,28 +164,6 @@ export default {
         ...mapGetters('courses', { courses: 'sortedCourses', retrievedAllCourses: 'retrievedAllCourses' }),
         ...mapGetters('user', { nameOfUser: 'name' }),
         ...mapGetters('pref', ['darkMode']),
-
-        // TODO: This is duplicated in Sidebar/CourseList.vue. We should factor
-        // it out into a Course or CourseCollection model or something.
-        courseExtraDataToDisplay() {
-            const getNameAndYear = c => `${c.name} (${c.created_at.slice(0, 4)})`;
-
-            const courseName = new Counter(this.courses.map(c => c.name));
-            const courseNameAndYear = new Counter(this.courses.map(getNameAndYear));
-
-            return this.courses.reduce((acc, course) => {
-                if (courseName.getCount(course.name) > 1) {
-                    if (courseNameAndYear.getCount(getNameAndYear(course)) > 1) {
-                        acc[course.id] = course.created_at.slice(0, 10);
-                    } else {
-                        acc[course.id] = course.created_at.slice(0, 4);
-                    }
-                } else {
-                    acc[course.id] = null;
-                }
-                return acc;
-            }, {});
-        },
 
         filteredCourses() {
             if (!this.searchString) {
@@ -319,14 +291,22 @@ export default {
         async showMoreCourses($state = null) {
             this.renderingMoreCourses += 1;
 
+            // This happens when you reload the courses using the sidebar, if we
+            // don't reset the amount to show we steadily increase the
+            // `amountCoursesToShow`, and at one point we will load all courses.
+            if (this.courses.length === 0) {
+                this.amountCoursesToShow = 0;
+            }
+
             const promises = [this.$afterRerender()];
             const nextToShow = this.amountCoursesToShow + EXTRA_COURSES_AMOUNT;
-            if (Math.min(nextToShow, this.courses.length) > INITIAL_COURSES_AMOUNT) {
+            if (nextToShow > INITIAL_COURSES_AMOUNT) {
                 promises.push(this.loadAllCourses());
             }
 
             await Promise.all(promises);
             this.amountCoursesToShow += EXTRA_COURSES_AMOUNT;
+            await this.$nextTick();
             this.renderingMoreCourses -= 1;
 
             if ($state) {
@@ -349,6 +329,7 @@ export default {
         LocalHeader,
         InfiniteLoading,
         AssignmentListItem,
+        CourseName,
     },
 };
 </script>