+++ src/components/Sidebar/index.vue
@@ -106,9 +106,7 @@
                         <icon :name="i ? 'arrow-left' : 'times'"/>
                     </div>
 
-                    <h4 class="submenu-header">
-                        {{ maybeCall(subMenu.header) }}
-                    </h4>
+                    <h4 class="submenu-header">{{ maybeCall(subMenu.header) }}</h4>
 
                     <div v-if="subMenu.reload || loading"
                          @click="refreshItems"
@@ -128,12 +126,12 @@
                 <hr class="separator"/>
 
                 <component :is="subMenu.component"
-                           v-show="subMenu.showWhenLoading || !loading"
-                           :data="maybeCall(subMenu.data)"
-                           @loading="loading = true"
-                           @loaded="loading = false"
-                           @open-menu="openSubMenu"
-                           @close-menu="closeSubMenu(true)"/>
+                            v-show="subMenu.showWhenLoading || !loading"
+                            :data="maybeCall(subMenu.data)"
+                            @loading="loading = true"
+                            @loaded="loading = false"
+                            @open-menu="openSubMenu"
+                            @close-menu="closeSubMenu(true)"/>
             </div>
         </div>
     </div>